refactor(ble): make separate module
